kzorp: Fixed packet drop in case of overlapping rules.
When matching rules with the greatest score have the same value
packet dropped because decision by the score cannot be made. Now
the best match is the rule with the lesser id, it solves the problem
and cause that the older rule remain working so the user will recognize
the overlapping rule.

References: jira ZORP-22